Add the GBM build.
The GBM build runs without X or Wayland and is more suitable for
standalone setups. This requires libinput and adding that and its
dependencies are most of this commit. IIUC libinput will be included in
the next version of the freedesktop SDK so this could (should?) be
removed when rebased to that.